requirements-3.11.txt
amqp==5.3.1
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.10.0
asttokens==3.0.0
async-lru==2.0.5
asyncssh==2.21.1
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
babel==2.17.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
billiard==4.2.2
bleach==6.3.0
cbor2==5.7.1
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.3
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.2.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.7
django-celery-beat==2.8.1
django-crontab==0.7.1
django-encrypted-model-fields==0.6.5
django-environ==0.12.0
django-filter==25.2
django-timezone-field==7.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
flower==2.0.1
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.45
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
invoke==2.2.1
ipykernel==7.1.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kombu==5.5.4
lark==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
msgpack==1.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==1.3.0
netmiko==4.6.0
notebook_shim==0.2.4
ntc_templates==8.1.0
numpy==2.3.4
openpyxl==3.1.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
paramiko==4.0.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
ply==3.11
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psutil==7.1.3
psycopg==3.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1>=0.6.1,<0.7.0
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodomex==3.23.0
Pygments==2.19.2
PyNaCl==1.6.0
pyOpenSSL==25.3.0
pyserial==3.5
pysmi==1.6.2
pysnmp==4.4.12
pytest==8.4.2
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.0.1
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scp==0.15.0
Send2Trash==1.8.3
service-identity==24.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
tacacs-plus==2.6
terminado==0.18.1
textfsm==2.1.0
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uri-template==1.3.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
zope.interface==8.1.1
